implemented Index<-->number with test
Index [0,max[ max is invalid
Number ]0,max] 0 is invalid

Index Number is a small feature that helps you have nice convertion helper when having code that uses index and number to reference the same data.

example if you have a vector of images you would normally iterate over the images using index so the first image have index 0, but client code could refer to that same image as image number 1.

In that Case I would use this helper class system to ensure that index and number is mixed and used wrongly.